% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicr.R
\name{.createReturnStructure}
\alias{.createReturnStructure}
\title{Create an s3 return structure for exported functions}
\usage{
.createReturnStructure(content, url, response)
}
\arguments{
\item{content}{Content of the response}

\item{url}{URL of the request}

\item{response}{httr response}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Create an s3 return structure for exported functions
}
