% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicr.R
\name{.handleFailure}
\alias{.handleFailure}
\title{Handle failure case for httr}
\usage{
.handleFailure(response)
}
\arguments{
\item{response}{httr error response}
}
\description{
Handle failure case for httr
}
