% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicr.R
\name{getAllAssertions}
\alias{getAllAssertions}
\title{Get a list of assertions}
\usage{
getAllAssertions(page = 1, count = 25)
}
\arguments{
\item{page}{the page number to retrieve}

\item{count}{the number of assertions to retrieve}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve all assertions from the CIViC DB
}
\examples{
getAllAssertions(count = 10)
getAllAssertions(page = 2, count = 10)
}
\keyword{assertions}
