% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicr.R
\name{getAllGenes}
\alias{getAllGenes}
\title{Get a list of genes}
\usage{
getAllGenes(page = 1, count = 25)
}
\arguments{
\item{page}{the page number to retrieve}

\item{count}{the number of genes to retrieve}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve all genes from the CIViC DB
}
\examples{
getAllGenes(count = 10)
getAllGenes(page = 2, count = 10)
}
\keyword{genes}
