% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicr.R
\name{getEvidenceItem}
\alias{getEvidenceItem}
\title{Get a specific evidence item}
\usage{
getEvidenceItem(id)
}
\arguments{
\item{id}{Internal CIViC ID of the evidence item of interest}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve a specific evidence item from the CIViC DB
}
\examples{
getEvidenceItem(id = 1)
}
\keyword{evidence}
\keyword{item}
