% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicr.R
\name{getVariant}
\alias{getVariant}
\title{Get a specific variant}
\usage{
getVariant(id)
}
\arguments{
\item{id}{Internal CIViC ID of the variant of interest}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve a specific variant from the CIViC DB
}
\examples{
getVariant(id = 1)
}
\keyword{variant}
